Add statistics ui

+++ core/model/src/main/java/hu/bme/aut/it9p0z/model/statistics/ConditionLogStatisticsModel.kt
@@ -1,9 +1,9 @@
 package hu.bme.aut.it9p0z.model.statistics
 
 data class ConditionLogStatisticsModel(
-    val feelings: HashMap<String,Int>,
-    val foodTriggers: HashMap<String,Int>,
-    val weatherTriggers: HashMap<String,Int>,
-    val mentalHealthTriggers: HashMap<String,Int>,
-    val otherTriggers: HashMap<String,Int>
+    val feelings: HashMap<String,Float>,
+    val foodTriggers: HashMap<String,Float>,
+    val weatherTriggers: HashMap<String,Float>,
+    val mentalHealthTriggers: HashMap<String,Float>,
+    val otherTriggers: HashMap<String,Float>
 )
